+++ src/ui/notification-manager.lua
@@ -1,7 +1,6 @@
 local ADDON_NAME, Addon = ...
 local AnchorFrame = Addon:GetModule("AnchorFrame")
 local NotificationManager = Addon:GetModule("NotificationManager")
-local SavedVariables = Addon:GetModule("SavedVariables")
 local Widgets = Addon:GetModule("Widgets")
 
 local activeNotifications = {}
@@ -43,20 +42,29 @@ function NotificationManager:Notify(text)
   table.insert(notificationQueue, 1, text)
 end
 
+function NotificationManager:NotifyWithIcon(icon, text)
+  self:Notify(Addon.TEXTURE_MESSAGE_FORMAT:format(icon, text))
+end
+
 -- ============================================================================
 -- Register callback to handle queued notifications.
 -- ============================================================================
 
 Addon:RegisterIntervalCallback(0.1, function()
-  local sv = SavedVariables:Get()
-  if #activeNotifications >= sv.maxNotifications then return end
+  local state = Addon:GetStore():GetState()
+  if #activeNotifications >= state.maxNotifications then return end
 
   local text = strtrim(table.remove(notificationQueue) or "")
   if text ~= "" then
     local notification = getNotification()
-    notification:Notify(text, sv.notificationFadeOutDelay)
-    notification:SetCallback(function() releaseNotification(notification) end)
     activeNotifications[#activeNotifications + 1] = notification
+    notification:Notify(
+      text,
+      state.notificationAlpha / Addon.NOTIFICATION_ALPHA_MAX,
+      state.notificationFadeOutDelay,
+      function()
+        releaseNotification(notification)
+      end)
   end
 end)
 
@@ -66,14 +74,15 @@ end)
 
 Addon:RegisterIntervalCallback(0.01, function()
   local points = AnchorFrame:GetNotificationPoints()
+  local spacing = AnchorFrame:GetNotificationSpacing(points)
   for i = #activeNotifications, 1, -1 do
     local notification = activeNotifications[i]
     notification:ClearAllPoints()
     if i == #activeNotifications then
       local relativePoint = AnchorFrame.frame:IsShown() and points.relativePoint or points.point
-      notification:SetPoint(points.point, AnchorFrame.frame, relativePoint, 0, 0)
+      notification:SetPoint(points.point, AnchorFrame.frame, relativePoint, 0, spacing)
     else
-      notification:SetPoint(points.point, activeNotifications[i + 1], points.relativePoint, 0, 0)
+      notification:SetPoint(points.point, activeNotifications[i + 1], points.relativePoint, 0, spacing)
     end
   end
 end)